Decrease offset size

+++ src/midi_fm.c
@@ -24,7 +24,7 @@ static MidiFmChannel fmChannels[MAX_FM_CHANS];
 static u8 pitchIsOutOfRange(u8 pitch);
 static u8 effectiveVolume(MidiFmChannel* channelState);
 static void updatePan(u8 chan);
-static u16 lookupFreqNum(u8 pitch, s16 offset);
+static u16 lookupFreqNum(u8 pitch, u8 offset);
 void midi_fm_reset(void);
 
 static const FmChannel** presets;
@@ -165,7 +165,7 @@ u8 midi_fm_pitchToOctave(u8 pitch)
     return (pitch - MIN_MIDI_PITCH) / SEMITONES;
 }
 
-static u16 lookupFreqNum(u8 pitch, s16 offset)
+static u16 lookupFreqNum(u8 pitch, u8 offset)
 {
     return FREQS[((u8)(pitch - MIN_MIDI_PITCH) % SEMITONES) + offset];
 }
